moved uploadingFile to context-menu component
