+++ client/src/app/app.module.ts
@@ -9,8 +9,12 @@ import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { DesktopComponent } from './desktop/desktop.component';
+
+//directives
 import { DesktopFullscreenDirective } from './desktop-fullscreen/desktop-fullscreen.directive';
-import { UploadingDirective } from './uploading/uploading.directive';
+import { UploadingDirective } from './context-menu/components/uploading/uploading.directive';
+import { AddNewFolderDirective } from './context-menu/components/addnewfolder/addnewfolder.directive';
+
 import { TaskbarComponent } from './taskbar/taskbar.component';
 import { FinderComponent } from './finder/finder.component';
 import { DrawComponent } from './draw/draw.component';
@@ -58,8 +62,8 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { DesktopModule } from './desktop/desktop.module';
 import { WidgetPanelComponent } from './widget-panel/widget-panel.component';
 import { ClockComponent } from './clock/clock.component';
-import { UploadingComponent } from './uploading/uploading.component';
-import { uploadingModule } from './uploading/uploading.module';
+import { UploadingComponent } from './context-menu/components/uploading/uploading.component';
+import { uploadingModule } from './context-menu/components/uploading/uploading.module';
 import { NotificationComponent } from './notification/notification.component';
 import { NotificationService } from './notification/notification.service';
 import { MobileComponent } from './mobile/mobile.component';
@@ -91,6 +95,7 @@ import {PersistanceService} from "./services/persistance.service";
 		RegisterComponent,
 		LogoutComponent,
 		UploadingDirective,
+		AddNewFolderDirective,
 		WidgetPanelComponent,
 		ClockComponent,
 		UploadingComponent,
