+++ client/src/app/context-menu/components/uploading/uploading.directive.ts
