+++ client/src/app/context-menu/components/uploading/uploading.module.ts
